refactor: Move all files to match their namespace
Also adds a test for the version module and remove an unused spec file.
