+++ lib/aspire_budget/core_extensions.rb
