+++ lib/aspire_budget/utils.rb
