+++ lib/aspire_budget/worksheets/backend_data.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative './worksheet_base'
+require 'aspire_budget/worksheets/worksheet_base'
 
 module AspireBudget
   module Worksheets
