--- spec/aspire_budget.spec.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe AspireBudget do
-  it 'has a version number' do
-    expect(AspireBudget::VERSION).not_to be nil
-  end
-
-  it 'does something useful' do
-    expect(false).to eq(true)
-  end
-end
