+++ spec/aspire_budget/version_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'aspire_budget/version'
+
+RSpec.describe AspireBudget do
+  it 'has a version number' do
+    expect(AspireBudget::VERSION).to be '0.0.1'
+  end
+end
