+++ spec/support/fixtures/v3-1-0/category_transfers_properties.json
