+++ spec/support/fixtures/v3-1-0/dashboard_properties.json
