+++ spec/support/fixtures/v3-1-0/transactions_properties.json
