+++ spec/support/fixtures/v3-2-0/category_transfers.json
