+++ spec/support/fixtures/v3-2-0/category_transfers_properties.json
