+++ spec/support/fixtures/v3-2-0/dashboard_properties.json
