+++ spec/support/fixtures/v3-2-0/transactions_properties.json
