+++ spec/support/fixtures/ws_dummy/dummy.json
